add hitpattern

+++ RecoTracker/FinalTrackSelectors/src/TrackListMerger.h
@@ -10,9 +10,9 @@
 // Original Author: Steve Wagner, stevew@pizero.colorado.edu
 // Created:         Sat Jan 14 22:00:00 UTC 2006
 //
-// $Author: wmtan $
-// $Date: 2011/06/20 20:28:26 $
-// $Revision: 1.5 $
+// $Author: stenson $
+// $Date: 2012/02/19 20:21:34 $
+// $Revision: 1.6 $
 //
 
 #include "FWCore/Framework/interface/EDProducer.h"
@@ -65,10 +65,10 @@ namespace cms
     double maxNormalizedChisq_;
     double minPT_;
     unsigned int minFound_;
-    double epsilon_;
-    double shareFrac_;
-    double foundHitBonus_;
-    double lostHitPenalty_;
+    float epsilon_;
+    float shareFrac_;
+    float foundHitBonus_;
+    float lostHitPenalty_;
 
     std::vector< std::vector< int> > listsToMerge_;
     std::vector<bool> promoteQuality_;
